Use RazorSlices
Migrate to use RazorSlices to maybe one day support publishing the app using AoT if it's ever supported by EFCore.
