+++ src/AppleFitnessWorkoutMapper/AppleFitnessWorkoutMapper.csproj
@@ -13,6 +13,7 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" />
     <PackageReference Include="Microsoft.Extensions.Resilience" />
     <PackageReference Include="Microsoft.TypeScript.MSBuild" PrivateAssets="all" />
+    <PackageReference Include="RazorSlices" />
   </ItemGroup>
   <ItemGroup>
     <Content Update="coverage\**\*;package.json;package-lock.json;tsconfig.json" CopyToPublishDirectory="Never" />
