+++ src/AppleFitnessWorkoutMapper/Slices/App.cshtml
@@ -1,8 +1,8 @@
-@page
-@using MartinCostello.AppleFitnessWorkoutMapper;
-@using MartinCostello.AppleFitnessWorkoutMapper.Pages;
-@addTagHelper "*, Microsoft.AspNetCore.Mvc.TagHelpers"
-@model IndexModel
+@inherits RazorSliceHttpResult<AppModel>
+@{
+    var ownerUrl = "https://github.com/martincostello";
+    var repoUrl = $"{ownerUrl}/apple-fitness-workout-mapper";
+}
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -12,7 +12,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-icons/1.8.3/font/bootstrap-icons.min.css" integrity="sha512-YzwGgFdO1NQw1CZkPoGyRkEnUTxPSbGWXvGiXrWk8IeSqdyci0dEDYdLLjMxq1zCoU0QBa4kHAFiRhUL3z2bow==" crossorigin="anonymous" referrerpolicy="no-referrer" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap4-toggle/3.6.1/bootstrap4-toggle.min.css" integrity="sha512-EzrsULyNzUc4xnMaqTrB4EpGvudqpetxG/WNjCpG6ZyyAGxeB6OBF9o246+mwx3l/9Cn838iLIcrxpPHTiygAA==" crossorigin="anonymous" referrerpolicy="no-referrer" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/css/bootstrap-datepicker.min.css" integrity="sha512-mSYUmp1HYZDFaVKK//63EcZq4iFWFjxSL+Z3T/aCt4IO9Cejm03q3NKKYN6pFQzY0SBOr8h+eCIAZHPXcpZaNw==" crossorigin="anonymous" referrerpolicy="no-referrer" />
-    <link rel="stylesheet" href="~/static/css/styles.css" asp-append-version="true">
+    <link rel="stylesheet" href="@(Content("~/static/css/styles.css"))">
     <link rel="dns-prefetch" href="https://cdnjs.cloudflare.com">
     <link rel="dns-prefetch" href="https://fonts.googleapis.com">
     <link rel="dns-prefetch" href="https://maps.googleapis.com">
@@ -27,8 +27,8 @@
 </head>
 <body>
     <nav class="navbar navbar-dark fixed-top bg-dark flex-md-nowrap p-0 shadow">
-        <a class="navbar-brand col-sm-3 col-md-2 mr-0" href="https://github.com/martincostello/apple-fitness-workout-mapper" rel="nofollow" target="_blank">
-            <img src="~/static/images/apple.svg" class="d-inline-block align-top" width="30" height="30" aria-hidden="true" asp-append-version="true" />
+        <a class="navbar-brand col-sm-3 col-md-2 mr-0" href="@(repoUrl)" rel="nofollow" target="_blank">
+            <img src="@(Content("~/static/images/apple.svg"))" class="d-inline-block align-top" width="30" height="30" aria-hidden="true" />
             Apple Fitness Workout Mapper
         </a>
         <form class="form-inline ml-auto">
@@ -51,7 +51,7 @@
                    data-size="sm">
             <ul class="navbar-nav">
                 <li class="nav-item">
-                    <a class="nav-link lead ml-2 mr-2 text-white" href="https://github.com/martincostello/apple-fitness-workout-mapper/blob/main/docs/help.md#help" rel="nofollow" target="_blank">
+                    <a class="nav-link lead ml-2 mr-2 text-white" href="@(repoUrl)/blob/main/docs/help.md#help" rel="nofollow" target="_blank">
                         <i class="bi-question-circle" aria-hidden="true"></i> Help
                     </a>
                 </li>
@@ -177,7 +177,7 @@
                 <div id="map" data-google-api-key="@(Model.GoogleMapsApiKey)"></div>
                 <footer class="mt-2 text-muted">
                     Designed with <i class="bi bi-heart-fill" aria-label="love" role="img"></i> by <a href="https://martincostello.com/home/about/" rel="nofollow" target="_blank">Martin Costello</a>, &copy; @DateTimeOffset.UtcNow.Year |
-                    <i class="bi bi-github" aria-hidden="true"></i> <a href="https://github.com/martincostello" rel="nofollow" target="_blank">GitHub</a> |
+                    <i class="bi bi-github" aria-hidden="true"></i> <a href="@(ownerUrl)" rel="nofollow" target="_blank">GitHub</a> |
                     <i class="bi bi-twitter" aria-hidden="true"></i> <a href="https://twitter.com/martin_costello" rel="nofollow" target="_blank">Twitter</a>
                 </footer>
             </main>
@@ -187,7 +187,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap/4.6.1/js/bootstrap.min.js" integrity="sha512-UR25UO94eTnCVwjbXozyeVd6ZqpaAE9naiEUBK/A+QDbfSTQFhPGj5lOR6d8tsgbBk84Ggb5A3EkjsOgPRPcKA==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap4-toggle/3.6.1/bootstrap4-toggle.min.js" integrity="sha512-bAjB1exAvX02w2izu+Oy4J96kEr1WOkG6nRRlCtOSQ0XujDtmAstq5ytbeIxZKuT9G+KzBmNq5d23D6bkGo8Kg==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/js/bootstrap-datepicker.min.js" integrity="sha512-T/tUfKSV1bihCnd+MxKD0Hm1uBBroVYBOYSk1knyvQ9VyZJpc/ALb4P0r6ubwVPSGB2GvjeoMAJJImBG12TiaQ==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-    <script src="~/static/js/main.js" asp-append-version="true"></script>
+    <script src="@(Content("~/static/js/main.js"))"></script>
     <script>
         $('.calendar').datepicker({
             autoclose: true,
@@ -201,9 +201,34 @@
     </script>
 </body>
 <!--
-    Build:         https://github.com/martincostello/apple-fitness-workout-mapper/actions/runs/@(GitMetadata.BuildId)
+    Build:         @(repoUrl)/actions/runs/@(GitMetadata.BuildId)
     Commit SHA:    @(GitMetadata.Commit)
     Commit branch: @(GitMetadata.Branch)
     Timestamp      @(GitMetadata.Timestamp)
 -->
 </html>
+
+@functions {
+    private string? Content(string contentPath)
+    {
+        string? result = string.Empty;
+
+        if (!string.IsNullOrEmpty(contentPath))
+        {
+            if (contentPath[0] == '~')
+            {
+                var segment = new PathString(contentPath[1..]);
+                var applicationPath = HttpContext!.Request.PathBase;
+
+                var path = applicationPath.Add(segment);
+                result = path.Value;
+            }
+            else
+            {
+                result = contentPath;
+            }
+        }
+
+        return $"{result}?v={GitMetadata.Commit}";
+    }
+}
