+++ src/AppleFitnessWorkoutMapper/Slices/_ViewImports.cshtml
@@ -1,0 +1,11 @@
+@inherits RazorSliceHttpResult
+
+@using System.Globalization;
+@using Microsoft.AspNetCore.Razor;
+@using Microsoft.AspNetCore.Http.HttpResults;
+@using MartinCostello.AppleFitnessWorkoutMapper;
+@using MartinCostello.AppleFitnessWorkoutMapper.Slices;
+@using RazorSlices
+
+@tagHelperPrefix __disable_tagHelpers__:
+@removeTagHelper *, Microsoft.AspNetCore.Mvc.Razor
